requirements.txt
selenium==4.1.3
pytest==7.1.2
termcolor==1.1.0
pytest-dependency==0.5.1
allure-pytest==2.9.45
paramiko~=2.11.0
